+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ Release Notes
 
 - Updated code to reduce warnings with latest ``numpy`` versions. [#16]
 
+- Optimized code to improve performance and minimize memory usage when either
+  ``masks`` and/or ``sigmas`` have default values. [#17]
+
 
 0.2.0 (07-August-2019)
 ======================